Update OS OPMPLC demo
This commit moves around some of the OPMPLC demo data and scripts. Furthermore, it adds a new command line flag:

`--enable-cors`: Switch on a dirty implementation of CORS on the web worker requests for tiles.

The new command line flag is useful if running behind something like HAProxy which can do the dirty work for you.